refactor: move custom catalog pages to ./src/pages
