+++ packages/app/src/App.tsx
@@ -22,7 +22,7 @@ import { TechDocsAddons } from '@backstage/plugin-techdocs-react';
 import { ReportIssue } from '@backstage/plugin-techdocs-module-addons-contrib';
 import { UserSettingsPage } from '@backstage/plugin-user-settings';
 import { apis } from './apis';
-import { entityPage } from './components/catalog/EntityPage';
+import { customEntityPage } from './pages/catalog/CustomEntityPage/CustomEntityPage';
 import { searchPage } from './components/search/SearchPage';
 import { Root } from './components/Root';
 import {
@@ -35,7 +35,7 @@ import { AppRouter, FlatRoutes } from '@backstage/core-app-api';
 import { CatalogGraphPage } from '@backstage/plugin-catalog-graph';
 import { RequirePermission } from '@backstage/plugin-permission-react';
 import { catalogEntityCreatePermission } from '@backstage/plugin-catalog-common/alpha';
-import { CustomCatalogIndexPage } from './pages/CustomCatalogIndexPage/CustomCatalogIndexPage';
+import { CustomCatalogIndexPage } from './pages/catalog/CustomCatalogIndexPage/CustomCatalogIndexPage';
 
 const app = createApp({
   apis,
@@ -71,7 +71,7 @@ const routes = (
       path="/catalog/:namespace/:kind/:name"
       element={<CatalogEntityPage />}
     >
-      {entityPage}
+      {customEntityPage}
     </Route>
     <Route path="/docs" element={<TechDocsIndexPage />} />
     <Route
