+++ packages/app/src/pages/catalog/CustomCatalogIndexPage/columns/index.ts
